fix download links, formatting, and removed dupes

+++ site/tmpl/editproblem/default.php
@@ -40,8 +40,8 @@ $uri = Uri::root();
 
 		$info = $this->details;
 		
-		$pdfExists = file_exists(dirname(__FILE__).'/../../../../media/com_catalogsystem/uploads/pdf/'.$info->pdfPath. '.pdf');
-		$zipExists = file_exists(dirname(__FILE__).'/../../../../media/com_catalogsystem/uploads/zip/'.$info->zipUrl. '.zip');
+		$pdfExists = file_exists(dirname(__FILE__).'/../../../../media/com_catalogsystem/uploads/pdf/'.$info->pdfPath);
+		$zipExists = file_exists(dirname(__FILE__).'/../../../../media/com_catalogsystem/uploads/zip/'.$info->zipUrl);
 		
 		// If a problem was edited successfully, display a temporary confirmation message
 		if($this->result->state === 0)
@@ -82,7 +82,7 @@ $uri = Uri::root();
 			}
             echo $this->form->renderFieldset("details");
 			
-			if($info->pdfPath != null){
+			if($info->pdfPath != null && $pdfExists){
 				$pdfDownload = $uri . "media/com_catalogsystem/uploads/pdf/" . $info->pdfPath;
 				echo "<div class= 'problem-header' style='display: flex;'><label id= 'pdf' class= 'upload-label'>Problem PDF:</label> <a class= 'title' href='$pdfDownload' target='_blank' rel='noopener noreferrer'>Download</a></div>";
 			} else {
@@ -93,11 +93,11 @@ $uri = Uri::root();
 			echo $this->form->renderField("pdfupload");
 			echo "</div>";
 			
-			if($info->zipUrl != null){
+			if($info->zipUrl != null && $zipExists){
 				$zipDownload = $uri . "media/com_catalogsystem/uploads/zip/" . $info->zipUrl;
 				echo "<div class= 'problem-header' style='display: flex;'><label class= 'upload-label' id= 'zip'>Problem ZIP:</label> <a class= 'title' href='$zipDownload' target='_blank' rel='noopener noreferrer' download>Download</a></div>";
 			} else {
-				echo "<div class= 'problem-header' style='display: flex;'><label class= 'upload-label' id= 'zip'>Problem ZIP:</label> <div class= 'title'>Not Available</div>";
+				echo "<div class= 'problem-header' style='display: flex;'><label class= 'upload-label' id= 'zip'>Problem ZIP:</label> <div class= 'title'>Not Available</div></div>";
 			}
 			
 			echo "<div class= 'problem-header'>";
